+++ js/data.js
@@ -1,0 +1,56 @@
+'use strict';
+
+const INITIAL_RANGE_LIKES = 15;
+const END_RANGE_LIKES = 200;
+const INITIAL_RANGE_COMMENTATOR_ID = 26;
+const END_RANGE_COMMENTATOR_ID = 200;
+const INITIAL_RANGE_AVATAR = 1;
+const END_RANGE_AVATAR = 6;
+
+const DESCRIPTIONS = [
+  'На прогулке',
+  'Ем травку',
+  'Бегаю по двору',
+  'Повис на занавесках',
+  'Утренняя зарядка',
+  'Вечерняя тренировка',
+  'Питаюсь правильно!',
+  'Играю с соседями',
+  'Греюсь на солнышке',
+  'Умываюсь',
+  'Шумлю в 3 ночи',
+  'Пытаюсь разбудить хозяев',
+  'Разбудил хозяев',
+  'Бегу от хозяев',
+  'Хозяева бегут за мной',
+  'Меня догнали',
+  'Кормят завтраком',
+  'Убирают лоток',
+  'Лежу на подоконнике',
+  'Лакаю молоко',
+  'Играем с друзьями в кости',
+  'Обедаю',
+  'Нашел сметану',
+  'Моюсь после сметаны',
+  'Сплю...',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Васёк',
+  'Антоша',
+  'Аня',
+  'Иван',
+  'Вика',
+  'Оля',
+  'Света'];
+  
+export {INITIAL_RANGE_LIKES, END_RANGE_LIKES, INITIAL_RANGE_COMMENTATOR_ID, END_RANGE_COMMENTATOR_ID, INITIAL_RANGE_AVATAR, END_RANGE_AVATAR, DESCRIPTIONS, MESSAGES, NAMES};
